+++ app/src/main/java/com/cajetan/youtubeplus/SeekDialog.java
@@ -3,9 +3,11 @@ package com.cajetan.youtubeplus;
 import android.app.Dialog;
 import android.app.DialogFragment;
 import android.app.TimePickerDialog;
+import android.content.Context;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
+import android.widget.Button;
 import android.widget.NumberPicker;
 
 public class SeekDialog extends DialogFragment {
@@ -15,6 +17,10 @@ public class SeekDialog extends DialogFragment {
     NumberPicker secondPicker;
     NumberPicker minutePicker;
     NumberPicker hourPicker;
+
+    Button confirmButton;
+
+    SeekDialogListener mListener;
 
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
@@ -38,7 +44,35 @@ public class SeekDialog extends DialogFragment {
         hourPicker.setMinValue(0);
         hourPicker.setMaxValue(100);
 
+        confirmButton = layoutView.findViewById(R.id.confirm_button);
+        confirmButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // TODO: maybe convert here
+
+                mListener.onSeekButtonClicked(hourPicker.getValue(), minutePicker.getValue(),
+                        secondPicker.getValue());
+
+                dismiss();
+            }
+        });
+
         return builder.create();
     }
 
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+
+        if (context instanceof SeekDialogListener) {
+            mListener = (SeekDialogListener) context;
+        } else {
+            throw new ClassCastException(getActivity().toString() +
+                    " must implement SeekDialogListener");
+        }
+    }
+
+    public interface SeekDialogListener {
+        void onSeekButtonClicked(int hours, int minutes, int seconds);
+    }
 }